+++ CHANGES.md
@@ -11,6 +11,7 @@
 - Unused storage plugins are loaded and cause non-fatal errors if dependencies are missing - `PR #799` - Thanks **electricworry**
 - Replaced usages of `asynctest` with `unittest.mock` in tests - `PR #807` and `PR #856` - Thanks **ichard26**
 - Remove debugging line that loads entire files into memory. - `PR #858` - Thanks **asrp**
+- Removed terrible isinstance check of unittest.Mock in mirror.py - `PR #859` - Thanks **ichard26**
 
 # 4.4.0 (2020-12-31)